Add files via upload
What is the use of %>% (piping) in R?
Pipes are an extremely useful tool from the magrittr package 1 that allows us to express a sequence of multiple operations. They can greatly simplify code and make operations more intuitive. However, they are not the only way to write your code and combine multiple operations.